chore: revert to python 3.13 (#662)
The app base image was bumped to Python 3.14 by dependabot and this was not reviewed carefully enough. It was also never active because we did not rebuild the purple-app-base image.

We seem to have a dependency on pydantic V1, probably through django-rest-framework, and pydantic V1 is incompatible with Python 3.14. Enabling tests using Python 3.14 revealed the warning about this.

Dropping back to Python 3.13, which functionally amounts to upgrading from 3.12 to 3.13, notwithstanding that the file previously said 3.14.