feat: working send_mail API (#743)
* chore: remove attachments from mail api

Was not working and is not required.

* feat: better email addr serialization

* feat: send email for real (wip)

Does not yet handle cc or celery-driven sending

* refactor: better sending email (still wip)

* feat: send messages via celery

* feat: retry mail delivery

* fix: json instead of multipart

The typescript client is using application/x-www-form-urlencoded for the multipart API. That can probably be made to work, but since we're not sending attachments, let's just use JSON.

* chore: log when sending email

* fix: drop AllowAny from mail APIs

* fix: fixup RfcMailTemplatesList

The serializer change for sending mail affected this. Fixing to accommodate.

* chore: remove unused import

* fix: typo in blank.txt template